refatorado teste quebrando

+++ stool_in_logic/lib/src/features/service_provider/domain/entity/service_provider/create_and_update_service_provider_entity.dart
@@ -5,7 +5,7 @@ class CreateAndUpdateServiceProviderEntity {
   final DateTime? initialDisponibleTime;
   final DateTime? endDisponibleTime;
   final String? disponibleDays;
-  final double? coins;
+  final int? coins;
   CreateAndUpdateServiceProviderEntity({
     this.userDescription,
     this.votes,
